refactor: Remove CNPJ field and add image upload API
The CNPJ field was removed from the barbershop schema as it is not a mandatory field and can be handled externally if needed.

An API function `uploadImage` has been added to facilitate uploading files to Supabase storage, which will be used for storing barbershop and user profile images. This enhances the platform's ability to manage visual assets.