fix: Enhance AppContext and payment handling
Updates the `AppContext` to include missing properties and modifies the `setPackageSubscriptionIntent` function to accept a `Barbershop` object instead of just its ID.

This change also refactors the `PaymentModal` component to properly handle Mercado Pago preferences and adds the `clientManagement` property to `IntegrationSettings`. Additionally, it improves the `create-mp-preference` API route by including the barbershop slug in the return URL and ensuring correct data fetching.